[Author: thakis]
Unset focus when hiding a per-tab sheet.

This is necessary so that when switching to a new tab using the
cmd-alt-right keyboard shortcut in chromium, the sheet in the previous
tab doesn't keep keyboard focus (which makes breaks cmd-l etc until the
window is clicked -- annoying).

R=thomasvl
DELTA=5  (5 added, 0 deleted, 0 changed)

git-svn-id: http://google-toolbox-for-mac.googlecode.com/svn@171 7dc7ac4e-7543-0410-b95c-c1676fc8e2a3