[Author: dmaclach]
There are still a couple of errors occurring, but this fixes up the majority of stuff. I figured I'd send it out to you now to take
a look at in the morning, and then we can clean up the last couple of issues.
Not quite sure what to do about the naming of data files so that we can encompass ranges of system version (e.g. this file applies to 
10.5 and 10.6 but not 10.4 or 10.7.

R=thomasvl
DELTA=10598  (10211 added, 70 deleted, 317 changed)

git-svn-id: http://google-toolbox-for-mac.googlecode.com/svn@339 7dc7ac4e-7543-0410-b95c-c1676fc8e2a3